corrections made on functions Bfx_SetBits and Bfx_PutBits suggested by daniel

+++ autosar/libs/Bfx_32bits.h
@@ -20,7 +20,7 @@ static inline boolean Bfx_GetBit_u32u8_u8( uint32 *Data, uint8 BitPn )
 
 static inline void Bfx_SetBits_u32u8u8u8( uint32 *Data, uint8 BitStartPn, uint8 BitLn, uint8 Status )
 {
-    uint8 Mask = ( ( 1u << BitLn ) - 1 ) << BitStartPn;
+    uint32 Mask = ( ( 1u << BitLn ) - 1 ) << BitStartPn;
 
     if( Status == 0 )
     {
@@ -134,7 +134,7 @@ static inline void Bfx_CopyBit_u32u8u32u8( uint32 *DestinationData, uint8 Destin
 
 static inline void Bfx_PutBits_u32u8u8u32( uint32 *Data, uint8 BitStartPn, uint8 BitLn, uint32 Pattern )
 {
-    uint8 Mask = ( ( 1u << BitLn ) - 1u ) << BitStartPn;
+    uint32 Mask = ( ( 1u << BitLn ) - 1u ) << BitStartPn;
     *Data      = ( *Data & ~Mask ) | ( ( Pattern << BitStartPn ) & Mask );
 }
